requirements.txt
telethon
python-dotenv
aiohttp
